Exercice [40]
The challenge 40 take advantage of a small e number (here 3). Using the
Chinese remainder theorem, it is possible to retrieve the secret message
using (e) encryptions result (ciphertext + pubkey). That's why it is
recommended to choose a high value for e (usually 65537) or even better,
to choose e randomly (but following some coprime constraints).