Add include headers for libcurl
On windows systems, curl is not installed by default or may not be visible
(PATH not updated) so we need to provide the includes headers used with
the static library used (of course you need to update those headers file
if you change the libcurl version).

Curl version used by the author : 7.40.0.